Feature/attention compatible fp16 fp32

+++ contrib/petr/petr_transformer.py
@@ -87,7 +87,7 @@ class PETRMultiheadFlashAttention(BaseModule):
         self.num_heads = num_heads
         self.batch_first = True
 
-        self.attn = FlashMHA(embed_dims, num_heads, attn_drop, dtype=torch.float16, device='cuda',
+        self.attn = FlashMHA(embed_dims, num_heads, attn_drop, device='cuda',
                                           **kwargs)
 
         self.proj_drop = nn.Dropout(proj_drop)
